done moving
